+++ cats/__init__.py
@@ -4,4 +4,4 @@ from .plugins import *
 from .types import *
 from .utils import *
 
-version = (4, 3, 0)
+version = (4, 3, 1)